+++ CHANGELOG.md
@@ -30,6 +30,8 @@ Changelog
 * add new interfaces `OSHDBTemporal` and `OSHDBBoundable` ([#369])
 * major improvements to code style guide adherence; fix some potential bugs found in static code analyis ([#374])
 * upgrade apache ignite to version 2.10.0 ([#386])
+* drop deprecated oshdb-api methods `where` and `osmTypes` ([#308])
+* clarify documentation of OSMContributionView results ([#292])
 * if cluster state is not active an exception will thrown ([#337])
 
 ### bugfixes
@@ -71,7 +73,9 @@ Changelog
   ```
 
 
+[#292]: https://github.com/GIScience/oshdb/issues/292
 [#306]: https://github.com/GIScience/oshdb/pull/306
+[#308]: https://github.com/GIScience/oshdb/issues/308
 [#327]: https://github.com/GIScience/oshdb/issues/327
 [#337]: https://github.com/GIScience/oshdb/issues/337
 [#338]: https://github.com/GIScience/oshdb/issues/338